useContent instead of params
