+++ client/App.js
@@ -1,12 +1,20 @@
-import AppNavigation from './src/navigation/appNavigation';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
+
+import AppNavigation from './src/navigation/appNavigation';
+import CurrentLocationProvider from './src/utils/CurrentLocationProvider';
+import RestaurantsProvider from './src/utils/RestaurantsProvider';
 
 export default function App() {
     return (
         <SafeAreaProvider>
-            <AppNavigation />
+            <RestaurantsProvider>
+                <CurrentLocationProvider>
+                    <AppNavigation />
+                </CurrentLocationProvider>
+            </RestaurantsProvider>
+            {/* <CurrentLocationProvider>
+                <AppNavigation />
+            </CurrentLocationProvider> */}
         </SafeAreaProvider>
     );
 }
-
-
