+++ client/src/screens/HomeScreen.js
@@ -1,20 +1,15 @@
-import React, { useState, createContext, useEffect } from 'react';
+import React, { useState, useContext } from 'react';
 import { SafeAreaView, View, Text, StyleSheet, TouchableOpacity, Image, FlatList } from 'react-native';
 import Swiper from 'react-native-swiper';
 import { icons, images, SIZES, COLORS } from '../constants';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
-export const currentLocationContext = createContext();
-export const restaurantContext = createContext();
+import { currentLocationContext, restaurantsContext } from '../utils/Context';
 
 const HomeScreen = ({ navigation }) => {
-    const initialCurrentLocation = {
-        streetName: 'Hanoi',
-        gps: {
-            latitude: 21.019835,
-            longitude: -254.215683,
-        },
-    };
+    // Use useContext instead of params
+    const initialCurrentLocation = useContext(currentLocationContext);
+    const restaurantData = useContext(restaurantsContext).restaurants;
 
     const categoryData = [
         {
@@ -69,165 +64,22 @@ const HomeScreen = ({ navigation }) => {
         },
     ];
 
-    // price rating
-    const affordable = 1;
-    const fairPrice = 2;
-    const expensive = 3;
-
-    const restaurantData = [
-        {
-            id: 1,
-            name: 'Burger',
-            rating: 4.8,
-            categories: [5, 7],
-            priceRating: affordable,
-            photo: images.burger_restaurant_1,
-            duration: '10 - 15 min',
-            location: {
-                latitude: 21.02154,
-                longitude: 105.77489,
-            },
-            courier: {
-                avatar: images.avatar_1,
-                name: 'Amy',
-            },
-            menu: [
-                {
-                    menuId: 1,
-                    name: 'Crispy Chicken Burger',
-                    photo: images.crispy_chicken_burger,
-                    description: 'Burger with crispy chicken, cheese and lettuce',
-                    calories: 200,
-                    price: 10,
-                    rating: 4.5,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-                {
-                    menuId: 2,
-                    name: 'Crispy Chicken Burger with Honey Mustard',
-                    photo: images.honey_mustard_chicken_burger,
-                    description: 'Crispy Chicken Burger with Honey Mustard Coleslaw',
-                    calories: 250,
-                    price: 15,
-                    rating: 4.5,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-            ],
-        },
-        {
-            id: 2,
-            name: 'Pizza',
-            rating: 4.8,
-            categories: [2, 4, 6],
-            priceRating: expensive,
-            photo: images.pizza_restaurant,
-            duration: '15 - 20 min',
-            location: {
-                latitude: 21.017459,
-                longitude: -254.215627,
-            },
-            courier: {
-                avatar: images.avatar_2,
-                name: 'Jackson',
-            },
-            menu: [
-                {
-                    menuId: 4,
-                    name: 'Hawaiian Pizza',
-                    photo: images.hawaiian_pizza,
-                    description: 'Canadian bacon, homemade pizza crust, pizza sauce',
-                    calories: 250,
-                    price: 15,
-                    rating: 4.5,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-                {
-                    menuId: 5,
-                    name: 'Tomato & Basil Pizza',
-                    photo: images.pizza,
-                    description: 'Fresh tomatoes, aromatic basil pesto and melted bocconcini',
-                    calories: 250,
-                    price: 20,
-                    rating: 4.5,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-                {
-                    menuId: 6,
-                    name: 'Tomato Pasta',
-                    photo: images.tomato_pasta,
-                    description: 'Pasta with fresh tomatoes',
-                    calories: 100,
-                    price: 10,
-                    rating: 4.5,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-                {
-                    menuId: 7,
-                    name: 'Mediterranean Chopped Salad ',
-                    photo: images.salad,
-                    description: 'Finely chopped lettuce, tomatoes, cucumbers',
-                    calories: 100,
-                    price: 10,
-                    rating: 4.8,
-                    comment: [
-                        {
-                            userId: 1,
-                            commentText: 'Delicious',
-                            userId: 2,
-                            commentText: 'Delicious',
-                        },
-                    ],
-                },
-            ],
-        },
-    ];
-
     const [categories, setCategories] = React.useState(categoryData);
     const [selectedCategory, setSelectedCategory] = React.useState(null);
     const [restaurants, setRestaurants] = React.useState(restaurantData);
+    const [idRestaurant, setIdRestaurant] = useState(null);
     const [currentLocation, setCurrentLocation] = React.useState(initialCurrentLocation);
     const [hasNotification, setHasNotification] = useState(false);
+
+    useContext(restaurantsContext).setIdRestaurant(idRestaurant);
 
     const insets = useSafeAreaInsets();
 
     function onSelectCategory(category) {
         //filter restaurant
         let restaurantList = restaurantData.filter((a) => a.categories.includes(category.id));
+
+        setRestaurants(restaurantList);
 
         setSelectedCategory(category);
     }
@@ -372,16 +224,15 @@ const HomeScreen = ({ navigation }) => {
             </View>
         );
     }
+
     function renderRestaurantSwiper() {
         const renderItem = ({ item }) => (
             <TouchableOpacity
-                style={{ marginBottom: SIZES.padding * 2 }}
-                onPress={() =>
-                    navigation.navigate('Restaurant', {
-                        item,
-                        currentLocation,
-                    })
-                }
+                style={{ marginBottom: SIZES.padding }}
+                onPress={() => {
+                    navigation.navigate('Restaurant');
+                    setIdRestaurant(item.id);
+                }}
             >
                 {/* Image */}
                 <View style={{ marginBottom: SIZES.padding }}>
@@ -394,23 +245,6 @@ const HomeScreen = ({ navigation }) => {
                             borderRadius: SIZES.radius,
                         }}
                     />
-
-                    <View
-                        style={{
-                            position: 'absolute',
-                            bottom: 0,
-                            height: 50,
-                            width: SIZES.width * 0.3,
-                            backgroundColor: COLORS.white,
-                            borderTopRightRadius: SIZES.radius,
-                            borderBottomLeftRadius: SIZES.radius,
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            ...styles.shadow,
-                        }}
-                    >
-                        <Text>{item.duration}</Text>
-                    </View>
                 </View>
 
                 {/* Restaurant Info */}
@@ -479,103 +313,103 @@ const HomeScreen = ({ navigation }) => {
     }
 
     function renderRestaurantList() {
-        const renderItem = ({ item }) => (
-            <TouchableOpacity
-                style={{ marginBottom: SIZES.padding * 2 }}
-                onPress={() =>
-                    navigation.navigate('Restaurant', {
-                        item,
-                        currentLocation,
-                    })
-                }
-            >
-                {/* Image */}
-                <View
-                    style={{
-                        marginBottom: SIZES.padding,
+        const renderItem = ({ item }) => {
+            return (
+                <TouchableOpacity
+                    style={{ marginBottom: SIZES.padding * 2 }}
+                    onPress={() => {
+                        navigation.navigate('Restaurant');
+                        setIdRestaurant(item.id);
                     }}
                 >
-                    <Image
-                        source={item.photo}
-                        resizeMode="cover"
-                        style={{
-                            width: '100%',
-                            height: 200,
-                            borderRadius: SIZES.radius,
-                        }}
-                    />
-
+                    {/* Image */}
                     <View
                         style={{
-                            position: 'absolute',
-                            bottom: 0,
-                            height: 50,
-                            width: SIZES.width * 0.3,
-                            backgroundColor: COLORS.white,
-                            borderTopRightRadius: SIZES.radius,
-                            borderBottomLeftRadius: SIZES.radius,
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            ...styles.shadow,
+                            marginBottom: SIZES.padding,
                         }}
                     >
-                        <Text>{item.duration}</Text>
+                        <Image
+                            source={item.photo}
+                            resizeMode="cover"
+                            style={{
+                                width: '100%',
+                                height: 200,
+                                borderRadius: SIZES.radius,
+                            }}
+                        />
+
+                        <View
+                            style={{
+                                position: 'absolute',
+                                bottom: 0,
+                                height: 50,
+                                width: SIZES.width * 0.3,
+                                backgroundColor: COLORS.white,
+                                borderTopRightRadius: SIZES.radius,
+                                borderBottomLeftRadius: SIZES.radius,
+                                alignItems: 'center',
+                                justifyContent: 'center',
+                                ...styles.shadow,
+                            }}
+                        >
+                            <Text>{item.duration}</Text>
+                        </View>
                     </View>
-                </View>
 
-                {/* Restaurant Info */}
-                <Text style={{ fontSize: 16, fontWeight: 'bold' }}>{item.name}</Text>
+                    {/* Restaurant Info */}
+                    <Text style={{ fontSize: 16, fontWeight: 'bold' }}>{item.name}</Text>
 
-                <View
-                    style={{
-                        marginTop: SIZES.padding,
-                        flexDirection: 'row',
-                    }}
-                >
-                    {/* Rating */}
-                    <Image
-                        source={icons.star}
-                        style={{
-                            height: 20,
-                            width: 20,
-                            tintColor: COLORS.primary,
-                            marginRight: 10,
-                        }}
-                    />
-                    <Text>{item.rating}</Text>
-
-                    {/* Categories */}
                     <View
                         style={{
+                            marginTop: SIZES.padding,
                             flexDirection: 'row',
-                            marginLeft: 10,
                         }}
                     >
-                        {item.categories.map((categoryId) => {
-                            return (
-                                <View style={{ flexDirection: 'row' }} key={categoryId}>
-                                    <Text>{getCategoryNameById(categoryId)}</Text>
-                                    <Text style={{ fontSize: 18, color: COLORS.darkgray }}> . </Text>
-                                </View>
-                            );
-                        })}
+                        {/* Rating */}
+                        <Image
+                            source={icons.star}
+                            style={{
+                                height: 20,
+                                width: 20,
+                                tintColor: COLORS.primary,
+                                marginRight: 10,
+                            }}
+                        />
+                        <Text>{item.rating}</Text>
 
-                        {/* Price */}
-                        {[1, 2, 3].map((priceRating) => (
-                            <Text
-                                key={priceRating}
-                                style={{
-                                    fontSize: 16,
-                                    color: priceRating <= item.priceRating ? COLORS.black : COLORS.darkgray,
-                                }}
-                            >
-                                $
-                            </Text>
-                        ))}
+                        {/* Categories */}
+                        <View
+                            style={{
+                                flexDirection: 'row',
+                                marginLeft: 10,
+                            }}
+                        >
+                            {item.categories.map((categoryId) => {
+                                return (
+                                    <View style={{ flexDirection: 'row' }} key={categoryId}>
+                                        <Text>{getCategoryNameById(categoryId)}</Text>
+                                        <Text style={{ fontSize: 18, color: COLORS.darkgray }}> . </Text>
+                                    </View>
+                                );
+                            })}
+
+                            {/* Price */}
+                            {[1, 2, 3].map((priceRating) => (
+                                <Text
+                                    key={priceRating}
+                                    style={{
+                                        fontSize: 16,
+                                        color: priceRating <= item.priceRating ? COLORS.black : COLORS.darkgray,
+                                    }}
+                                >
+                                    $
+                                </Text>
+                            ))}
+                        </View>
                     </View>
-                </View>
-            </TouchableOpacity>
-        );
+                </TouchableOpacity>
+            );
+        };
 
         return (
             <FlatList
@@ -593,14 +427,12 @@ const HomeScreen = ({ navigation }) => {
     }
 
     return (
-        <currentLocationContext.Provider value={currentLocation}>
-            <SafeAreaView style={[styles.container, { paddingTop: insets.top, paddingBottom: 50 }]}>
-                {renderHeader()}
-                {renderMainCategories()}
-                {renderRestaurantSwiper()}
-                {renderRestaurantList()}
-            </SafeAreaView>
-        </currentLocationContext.Provider>
+        <SafeAreaView style={[styles.container, { paddingTop: insets.top, paddingBottom: 50 }]}>
+            {renderHeader()}
+            {renderMainCategories()}
+            {renderRestaurantSwiper()}
+            {renderRestaurantList()}
+        </SafeAreaView>
     );
 };
