+++ client/src/utils/Context.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+export const currentLocationContext = createContext();
+export const restaurantsContext = createContext();
